requirements.txt
  
fastapi
sqlalchemy
psycopg2
pydantic
python-dotenv
uvicorn
